Fixed a bunch of Clang static analyzer issues.
As part of this, -[ShipEntity newMissile] was renamed to -selectMissile, because the "new" prefix conventionally indicates a method that returns a new object which is not autoreleased (as in -[Universe newShipWithRole:]).

Also introduced a bunch of annoying but technically correct "potential leak" issues. (These all refer to cases where an object is weakRetained and passed to JS_SetPrivate(). They aren't actual leaks, but they don't follow normal OpenStep retain/release behaviour.)"

git-svn-id: http://svn.berlios.de/svnroot/repos/oolite-linux/trunk@2996 127b21dd-08f5-0310-b4b7-95ae10353056