Primable equipment work - add affinity parameters for equipment.plist
fast_affinity_defensive, fast_affinity_offensive, both boolean. If set and that fast activation slot (defensive = 0, offensive = tab) has nothing in it (or what it had in it is no longer installed), assign that slot. As suggested in comments on db8f3ed

More cleanup of old energy bomb code.