Create standalone atmosphere texture generator for planets
Still seems to be some white-ball bugs when using the PlanetTexture generator and this will also be more efficient.

Still to do - but perhaps wait and see if this has fixed the bug - is to move some of the static functions that this has in common with PlanetTexture generator into a more sensible place.