Adjust the boundary between the two render passes
Adjusting the boundary to 100m means that most of the rendering takes place with a near clip plane of 100, which gives acceptable results on a 24-bit depth buffer at visual ranges.

There's still a need to do a short-range rendering pass for the few objects in that region, but it's fairly cheap.