Allow OXPs to be restricted to only be loaded during a scenario
And not during default play - so they will only be loaded if they are
a scenario tag or identifier, or are required by an OXP that is.

Allows installing "major conversion" style OXPs for an alternative
scenario, without those OXPs being loaded in normal play (so you
could even have more than one at once installed)

Slight catch: the scenario itself can't be in that OXP - if it is, then
it won't be on the "start new game" screen because it won't be loaded at
that point. This is easy enough to work around.