Add 'i' for information to OXZ manager
This allows additional paragraphs to be put into the description tag.
Before this is merged and people start being encouraged to use that for
extra readme information, the manifest API on the website needs to be
modified to only send the first paragraph of the description to Oolite/1.80
or earlier.