In "extra detail" mode, use even higher resolution planet textures.
Less blurry look from orbit, and seems to generate in acceptable time.
4096x4096 scale works fine with a modern graphics card too, and looks even
better ... but we'd need to get pre-generation of planet textures working
first to avoid unpleasant pauses when exiting hyperspace.

This is slightly experimental - it works for me but needs testing on other
high-end (by Oolite standards...) systems as well.