Windows Resources Addition & Manifest Embedding (#279)
Added resource files and enabled manifest embedding on Windows. This allows us to handle high-DPI and newer Windows versions support in one go (and is the recommended by Microsoft way for handling these things), can provide a way for potential future manifest settings to be added, simplifies existing code and adds an icon to the Oolite Windows executable.