Optimization: Explosion Culling (#292)
You may have noticed already that, sometimes, entity count jumps quite a bit, then drops quickly back down to its previous levels. This is often due to explosions. They tend to generate a number of entities that are added to the universe at the moment of the explosion and are removed shortly after, once their life has expired. This happens every time as part of the simulation, regardless of where the explosion takes place. Very often, the amount of entities being generated results in momentary stutter.

This change tries to minimize the side effects of explosions happening. We now generate all those extra entities only if the object that exploded was visible in the first place. Regardless of visibility, the explosion is rendered normally if it has happened within scanner range (so e.g. an escape pod exploding will always have a visible explosion, although it stops being visible past 5000m or so).