The Javascript runtime size can now be adjusted.
Use the key 'JSRunTime_size_mb' in .GNUstepDefaults. The value of this key is the requested size of the JS runtime in megabytes. If the key is missing, then the default value of 32MB is used. This might be useful in the case of a lot of script-heavy (i.e. memory consuming) OXPs running.