Added HDR display device detection capability on Windows.
If -hdr has been passed as startup parameter, we now check whether the primary display (which is where Oolite starts on) is HDR enabled and notify the user in case it is not. If no HDR display capable primary device has been found, the user has the choice to launch the game anyway or quit.