Improved HDR detection.
We are now using more reliable incoming Windows 11 24H2 APIs for HDR detection as a first choice. If this method fails (and it is expected to fail if Win11 24H2 official or preview is not installed), we fall back to the legacy detection method.
This commit requires the 20240929 or later version of the Oolite Windows Development Environment.