Implemented new ECM visual effect.
An EMP disruption effect similar to a burst of TV static while an ECM blast is active has been implemented. The effect will occur regardless of who is triggering the ECM, be it player or NPC. It can be disabled by setting the key "ecm-visual-fx" to NO in the user preferences file.