Pause in mission screens (#518)
Pause in mission screens is now allowed when docked

There does not seem to be a reason why pause / unpause cannot be used in mission screens. Unpause should work even if the mission screen is in text entry mode for some reason. Note that this is for docked mission screens only, in-flight ones are not considered right now.